# openapi.yaml
openapi: 3.0.3
info:
  title: GALO - API de datos de clientes (proveedor externo)
  description: >
    API para consumo interno de la informacion de clientes obtenida del proveedor externo,
    con enmascaramiento de campos sensibles y revelacion auditada bajo autorizacion.
  version: "1.0.0"
servers:
  - url: /
paths:
  /health:
    get:
      summary: Estado del servicio y de la base de datos
      security: []
      responses:
        "200":
          description: OK
  /auth/login:
    post:
      summary: Autenticacion de usuarios (dashboard / equipos consumidores humanos)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200": { description: Token JWT emitido }
        "401": { description: Credenciales invalidas }
  /api/v1/usuarios:
    get:
      summary: Lista paginada de clientes con campos sensibles enmascarados
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: query
          name: page
          schema: { type: integer, default: 1 }
        - in: query
          name: pageSize
          schema: { type: integer, default: 20 }
        - in: query
          name: search
          schema: { type: string }
      responses:
        "200": { description: Listado paginado }
        "401": { description: No autenticado }
  /api/v1/usuarios/{id}:
    get:
      summary: Detalle de un cliente (campos sensibles enmascarados)
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200": { description: OK }
        "404": { description: No encontrado }
  /api/v1/usuarios/{id}/reveal:
    post:
      summary: Revela los campos sensibles en claro (solo admin, requiere justificacion y queda auditado)
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 10 }
      responses:
        "200": { description: Datos revelados }
        "403": { description: Rol sin permisos }
        "429": { description: Limite de revelaciones alcanzado }
  /api/v1/ingestion/status:
    get:
      summary: Historial de las ultimas corridas de ingesta
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "200": { description: OK }
  /api/v1/ingestion/run:
    post:
      summary: Dispara manualmente una corrida de ingesta del proveedor externo
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        "200": { description: Resultado de la ingesta }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key